Show a Profile's status on the Profile information pane
There is a segfault that occurs when AppAnvil exits
