+++ src/tabs/model/database.cc
@@ -1,5 +1,6 @@
 #include "database.h"
 #include <exception>
+#include <gtkmm/treeiter.h>
 #include <memory>
 #include <stdexcept>
 
@@ -42,15 +43,26 @@ void Database::put_log_data(const std::string profile_name, const unsigned int p
 
 void Database::register_profile_column_record(std::shared_ptr<StatusColumnRecord> col_record)
 {
-    if(col_record == NULL) {
+    if(col_record == nullptr) {
         throw new std::runtime_error("Error: Argument for ColumnRecord (Profile) is null.");
-    } else if (prof_col_record != NULL) {
+    } else if (prof_col_record != nullptr) {
         throw new std::runtime_error("Error: Attempted to initialize ColumnRecord (Profile) more than once.");
     }
 
     prof_col_record = col_record;
 }
 
+ProfileTableEntry Database::get_profile_data(const std::string &profile_name){
+    auto iter = profile_data.find(profile_name);
+
+    if(iter == profile_data.end()){
+        // No data was found, so make up a fake entry
+        return ProfileTableEntry(profile_name, UNKNOWN_STATUS, Gtk::TreeRow());;
+    }
+    
+    // Return the found entry
+    return iter->second;
+}
 
 Database::Database()
 {   }
